has problem, stack smashing detected

+++ new/visualize.cpp
@@ -39,7 +39,7 @@ int main(int argc, char** argv) {
     glClearColor( 0.0f, 0.0f, 0.0f, 1.0f );
 
     int numberOfParticles = 30;
-    Position* position = init(numberOfParticles, WIDTH, HEIGHT);
+    Position* position = init((int)numberOfParticles, (int)WIDTH, (int)HEIGHT);
 
     const GLfloat point[] = {0.0f, 0.0f, 0.0f};
     GLuint vbo;
